// src/interfaces/post.ts
import { type Author } from "./author";

export type Post = {
  slug: string;
  title: string;
  date: string;
  coverImage: string;
  author: Author;
  excerpt: string;
  ogImage: {
    url: string;
  };
  content: string;
  preview?: boolean;
  tags?: string[];
  readingTime?: {
    minutes: number;
    text: string;
  };
  wordCount?: number;
};